requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-genai
pydantic
python-dotenv
